requirements.txt
aiohttp==1.0.5
async-timeout==2.0.0
Brlapi==0.6.7
certifi==2018.1.18
chardet==3.0.4
cupshelpers==1.0
discord.py==0.16.12
Flask==0.10.1
httplib2==0.9.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
louis==3.4.1
MarkupSafe==1.0
multidict==4.1.0
oauthlib==2.0.6
olefile==0.45.1
pew==1.1.2
pexpect==4.2.1
Pillow==4.3.0
pipenv==10.0.0
pycairo==1.15.4
pycups==1.9.73
pycurl==7.43.0.1
pygame==1.9.3
pygobject==3.26.1
pypubg==0.1.5
pysmbc==1.0.15.6
python-apt==1.4.0b3+b1
python-debian==0.1.32
python-debianbts==2.7.2
pyxdg==0.25
reportbug==7.1.8
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==0.5.0
six==1.11.0
unattended-upgrades==0.1
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
websockets==3.4
Werkzeug==0.14.1
